+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [unreleased]
 
 - Added [pwmake][] Comments
+- Changed [Set-DotEnv][] Read `.env` to `.myenv`
 
 ## [0.9.0] - 2025-05-31